@@ -4,7 +4,8 @@ import com.mongodb.reactivestreams.client.MongoClient;
 import com.mongodb.reactivestreams.client.MongoClients;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang.StringUtils;
+
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.boot.autoconfigure.mongo.MongoProperties;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
@@ -29,9 +30,8 @@ public class MongoDbConfiguration extends AbstractReactiveMongoConfiguration {
   @Override
   public MongoClient reactiveMongoClient() {
     log.info("Creating Mongo client with {}...", mongoProperties.getUri());
-    return StringUtils.isBlank(mongoProperties.getUri()) ?
-        MongoClients.create() :
-        MongoClients.create(mongoProperties.getUri());
+    return StringUtils.isBlank(mongoProperties.getUri()) ? MongoClients.create()
+        : MongoClients.create(mongoProperties.getUri());
   }
 
   @Override
